requirements.txt
asgiref==3.5.2
Django==4.0.6
django-admin-thumbnails==0.2.6
django-phonenumber-field==6.3.0
djangorestframework==3.13.1
phonenumbers==8.12.51
Pillow==9.2.0
python-decouple==3.6
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
